{"schema":"maestro/commit/1","summary":"extract overlay frozen selection and brush runtime modules","authority":"XY-242"}
