+++ packages/rsnap-overlay/src/overlay.rs
@@ -6,6 +6,8 @@ mod capture_window_runtime;
 mod config_runtime;
 mod cursor_context_runtime;
 mod cursor_runtime;
+mod frozen_brush_runtime;
+mod frozen_selection_runtime;
 mod hud_helpers;
 mod hud_runtime;
 mod image_helpers;
@@ -1331,1802 +1333,6 @@ impl OverlaySession {
 				self.hide_capture_windows();
 			}
 		}
-	}
-
-	fn update_live_drag_rect(&mut self, monitor: MonitorRect, global: GlobalPoint) {
-		let was_active = self.live_drag_hides_auxiliary_windows();
-
-		if !matches!(self.state.mode, OverlayMode::Live) {
-			self.state.drag_rect = None;
-
-			return;
-		}
-		if !self.left_mouse_button_down || self.left_mouse_button_down_monitor != Some(monitor) {
-			self.state.drag_rect = None;
-
-			return;
-		}
-
-		let Some(start_global) = self.left_mouse_button_down_global else {
-			self.state.drag_rect = None;
-
-			return;
-		};
-		let Some(rect) = monitor.local_rect_from_points(start_global, global) else {
-			self.state.drag_rect = None;
-
-			return;
-		};
-
-		if rect.is_empty() {
-			self.state.drag_rect = None;
-
-			return;
-		}
-
-		self.state.drag_rect = Some(MonitorRectPoints { monitor_id: monitor.id, rect });
-
-		if !was_active {
-			self.hide_auxiliary_windows_for_live_drag();
-		}
-	}
-
-	fn frozen_capture_rect_drag_target(&self) -> Option<(MonitorRect, RectPoints)> {
-		if !matches!(self.state.mode, OverlayMode::Frozen)
-			|| self.frozen_capture_source != FrozenCaptureSource::DragRegion
-			|| self.scroll_capture.active
-			|| self.state.frozen_image.is_none()
-		{
-			return None;
-		}
-
-		let monitor = self.state.monitor?;
-		let capture_rect = self.state.frozen_capture_rect?;
-
-		if capture_rect.is_empty() {
-			return None;
-		}
-
-		Some((monitor, capture_rect))
-	}
-
-	fn frozen_selection_drag_target(&self) -> Option<(MonitorRect, RectPoints)> {
-		(self.toolbar_state.selected_tool == FrozenToolbarTool::Pointer)
-			.then(|| self.frozen_capture_rect_drag_target())
-			.flatten()
-	}
-
-	fn frozen_capture_rect_for_monitor(&self, monitor: MonitorRect) -> Option<RectPoints> {
-		self.state
-			.frozen_capture_rect
-			.or_else(|| Some(RectPoints::new(0, 0, monitor.width, monitor.height)))
-			.filter(|capture_rect| !capture_rect.is_empty())
-	}
-
-	fn frozen_brush_capture_target(&self) -> Option<(MonitorRect, RectPoints)> {
-		if !matches!(self.state.mode, OverlayMode::Frozen)
-			|| self.scroll_capture.active
-			|| self.state.frozen_image.is_none()
-		{
-			return None;
-		}
-
-		let monitor = self.state.monitor?;
-		let capture_rect = self.frozen_capture_rect_for_monitor(monitor)?;
-
-		Some((monitor, capture_rect))
-	}
-
-	fn frozen_auto_center_available(&self) -> bool {
-		self.frozen_capture_rect_drag_target().is_some()
-	}
-
-	fn frozen_mosaic_drag_target(&self) -> Option<(MonitorRect, RectPoints)> {
-		if !matches!(self.state.mode, OverlayMode::Frozen)
-			|| self.scroll_capture.active
-			|| self.state.frozen_image.is_none()
-			|| !self.frozen_final_capture_ready()
-			|| self.toolbar_state.selected_tool != FrozenToolbarTool::Mosaic
-		{
-			return None;
-		}
-
-		let monitor = self.state.monitor?;
-		let capture_rect = self.state.frozen_capture_rect?;
-
-		if capture_rect.is_empty() {
-			return None;
-		}
-
-		Some((monitor, capture_rect))
-	}
-
-	fn begin_frozen_selection_drag(&mut self, global: GlobalPoint) -> bool {
-		let Some((monitor, capture_rect)) = self.frozen_selection_drag_target() else {
-			return false;
-		};
-		let Some((cursor_x, cursor_y)) = monitor.local_u32(global) else {
-			return false;
-		};
-		let Some(interaction) =
-			Self::frozen_selection_interaction_kind(capture_rect, cursor_x, cursor_y)
-		else {
-			return false;
-		};
-
-		self.frozen_selection_drag = FrozenSelectionDragState {
-			active: true,
-			interaction,
-			anchor_rect: capture_rect,
-			pointer_offset_x: cursor_x.saturating_sub(capture_rect.x),
-			pointer_offset_y: cursor_y.saturating_sub(capture_rect.y),
-			press_cursor_x: cursor_x,
-			press_cursor_y: cursor_y,
-		};
-
-		self.hide_auxiliary_windows_for_frozen_selection_drag();
-
-		true
-	}
-
-	fn begin_frozen_mosaic_drag(&mut self, global: GlobalPoint) -> bool {
-		let Some((monitor, capture_rect)) = self.frozen_mosaic_drag_target() else {
-			return false;
-		};
-		let Some((cursor_x, cursor_y)) = monitor.local_u32(global) else {
-			return false;
-		};
-
-		if !capture_rect.contains((cursor_x, cursor_y)) {
-			return false;
-		}
-
-		self.frozen_mosaic_drag =
-			FrozenMosaicDragState { active: true, anchor_x: cursor_x, anchor_y: cursor_y };
-		self.state.frozen_mosaic_preview_rect = Some(RectPoints::new(cursor_x, cursor_y, 1, 1));
-
-		true
-	}
-
-	fn frozen_selection_interaction_kind(
-		capture_rect: RectPoints,
-		cursor_x: u32,
-		cursor_y: u32,
-	) -> Option<FrozenSelectionInteractionKind> {
-		if let Some(corner) = WindowRenderer::frozen_selection_resize_hit_test(
-			capture_rect,
-			Pos2::new(cursor_x as f32, cursor_y as f32),
-		) {
-			return Some(FrozenSelectionInteractionKind::Resize(corner));
-		}
-
-		capture_rect.contains((cursor_x, cursor_y)).then_some(FrozenSelectionInteractionKind::Move)
-	}
-
-	fn frozen_selection_resize_cursor_icon(corner: FrozenSelectionCorner) -> CursorIcon {
-		match corner {
-			FrozenSelectionCorner::TopLeft | FrozenSelectionCorner::BottomRight => {
-				CursorIcon::NwseResize
-			},
-			FrozenSelectionCorner::TopRight | FrozenSelectionCorner::BottomLeft => {
-				CursorIcon::NeswResize
-			},
-		}
-	}
-
-	#[cfg_attr(target_os = "macos", allow(dead_code))]
-	fn frozen_selection_cursor_icon_for_monitor(&self, monitor: MonitorRect) -> CursorIcon {
-		if let Some((target_monitor, capture_rect)) = self.frozen_mosaic_drag_target() {
-			if target_monitor != monitor {
-				return CursorIcon::Default;
-			}
-			if self.frozen_mosaic_drag.active {
-				return CursorIcon::Crosshair;
-			}
-
-			let Some(cursor) = self.state.cursor else {
-				return CursorIcon::Default;
-			};
-			let Some((cursor_x, cursor_y)) = monitor.local_u32(cursor) else {
-				return CursorIcon::Default;
-			};
-
-			return if capture_rect.contains((cursor_x, cursor_y)) {
-				CursorIcon::Crosshair
-			} else {
-				CursorIcon::Default
-			};
-		}
-
-		if self.toolbar_state.selected_tool == FrozenToolbarTool::Pen
-			&& let Some((target_monitor, capture_rect)) = self.frozen_brush_capture_target()
-		{
-			if target_monitor != monitor {
-				return CursorIcon::Default;
-			}
-			if self.frozen_brush.active_stroke.is_some() {
-				return CursorIcon::Crosshair;
-			}
-
-			let Some(cursor) = self.state.cursor else {
-				return CursorIcon::Default;
-			};
-			let Some((cursor_x, cursor_y)) = monitor.local_u32(cursor) else {
-				return CursorIcon::Default;
-			};
-
-			return if capture_rect.contains((cursor_x, cursor_y)) {
-				CursorIcon::Crosshair
-			} else {
-				CursorIcon::Default
-			};
-		}
-
-		let Some((target_monitor, capture_rect)) = self.frozen_selection_drag_target() else {
-			return CursorIcon::Default;
-		};
-
-		if target_monitor != monitor {
-			return CursorIcon::Default;
-		}
-		if self.frozen_selection_drag.active {
-			return match self.frozen_selection_drag.interaction {
-				FrozenSelectionInteractionKind::Resize(corner) => {
-					Self::frozen_selection_resize_cursor_icon(corner)
-				},
-				FrozenSelectionInteractionKind::Move => CursorIcon::Grabbing,
-			};
-		}
-
-		let Some(cursor) = self.state.cursor else {
-			return CursorIcon::Default;
-		};
-		let Some((cursor_x, cursor_y)) = monitor.local_u32(cursor) else {
-			return CursorIcon::Default;
-		};
-		let cursor_local = Pos2::new(cursor_x as f32, cursor_y as f32);
-
-		if let Some(hit_rect) = self.frozen_text_edit_hit_rect_for_monitor(monitor)
-			&& hit_rect.contains(cursor_local)
-		{
-			return if self.frozen_text_edit.as_ref().is_some_and(|edit| edit.dragging) {
-				CursorIcon::Grabbing
-			} else {
-				CursorIcon::Grab
-			};
-		}
-
-		if self.frozen_text_tool_active() && capture_rect.contains((cursor_x, cursor_y)) {
-			return CursorIcon::Text;
-		}
-
-		match Self::frozen_selection_interaction_kind(capture_rect, cursor_x, cursor_y) {
-			Some(FrozenSelectionInteractionKind::Resize(corner)) => {
-				Self::frozen_selection_resize_cursor_icon(corner)
-			},
-			Some(FrozenSelectionInteractionKind::Move) => CursorIcon::Grab,
-			None => CursorIcon::Default,
-		}
-	}
-
-	#[cfg_attr(target_os = "macos", allow(dead_code))]
-	fn overlay_cursor_icon_for_monitor(&self, monitor: MonitorRect) -> CursorIcon {
-		match self.state.mode {
-			OverlayMode::Frozen => self.frozen_selection_cursor_icon_for_monitor(monitor),
-			OverlayMode::Live => CursorIcon::Crosshair,
-		}
-	}
-
-	#[cfg(target_os = "macos")]
-	fn frozen_selection_cursor_rects_for_monitor(
-		&self,
-		monitor: MonitorRect,
-	) -> Vec<OverlayCursorRect> {
-		let overlay_bounds =
-			Rect::from_min_size(Pos2::ZERO, Vec2::new(monitor.width as f32, monitor.height as f32));
-
-		if matches!(self.state.mode, OverlayMode::Live) {
-			return vec![OverlayCursorRect::new(overlay_bounds, CursorIcon::Crosshair)];
-		}
-		if !matches!(self.state.mode, OverlayMode::Frozen) {
-			return Vec::new();
-		}
-
-		if let Some((target_monitor, capture_rect)) = self.frozen_mosaic_drag_target() {
-			if target_monitor != monitor {
-				return Vec::new();
-			}
-			if self.frozen_mosaic_drag.active {
-				return vec![OverlayCursorRect::new(overlay_bounds, CursorIcon::Crosshair)];
-			}
-
-			let capture_rect = Rect::from_min_size(
-				Pos2::new(capture_rect.x as f32, capture_rect.y as f32),
-				Vec2::new(capture_rect.width as f32, capture_rect.height as f32),
-			)
-			.intersect(overlay_bounds);
-
-			return (capture_rect.width() > 0.0 && capture_rect.height() > 0.0)
-				.then_some(vec![OverlayCursorRect::new(capture_rect, CursorIcon::Crosshair)])
-				.unwrap_or_default();
-		}
-
-		let Some((target_monitor, capture_rect)) = self.frozen_selection_drag_target() else {
-			return Vec::new();
-		};
-
-		if target_monitor != monitor {
-			return Vec::new();
-		}
-		if self.frozen_selection_drag.active {
-			return match self.frozen_selection_drag.interaction {
-				FrozenSelectionInteractionKind::Resize(corner) => vec![OverlayCursorRect::new(
-					overlay_bounds,
-					Self::frozen_selection_resize_cursor_icon(corner),
-				)],
-				FrozenSelectionInteractionKind::Move => {
-					vec![OverlayCursorRect::new(overlay_bounds, CursorIcon::Grabbing)]
-				},
-			};
-		}
-
-		Self::frozen_selection_hover_cursor_rects(capture_rect)
-			.into_iter()
-			.filter_map(|cursor_rect| {
-				let rect = cursor_rect.rect.intersect(overlay_bounds);
-
-				(rect.width() > 0.0 && rect.height() > 0.0)
-					.then_some(OverlayCursorRect::new(rect, cursor_rect.icon))
-			})
-			.collect()
-	}
-
-	#[cfg(target_os = "macos")]
-	fn frozen_selection_hover_cursor_rects(capture_rect: RectPoints) -> Vec<OverlayCursorRect> {
-		let selection_rect = Rect::from_min_size(
-			Pos2::new(capture_rect.x as f32, capture_rect.y as f32),
-			Vec2::new(capture_rect.width as f32, capture_rect.height as f32),
-		);
-		let interior_reach_x = (selection_rect.width() * 0.35)
-			.min(FROZEN_SELECTION_RESIZE_HANDLE_INTERIOR_REACH_MAX_POINTS);
-		let interior_reach_y = (selection_rect.height() * 0.35)
-			.min(FROZEN_SELECTION_RESIZE_HANDLE_INTERIOR_REACH_MAX_POINTS);
-		let mut x_edges = vec![
-			selection_rect.min.x,
-			selection_rect.min.x + interior_reach_x,
-			selection_rect.center().x,
-			selection_rect.max.x - interior_reach_x,
-			selection_rect.max.x,
-		];
-		let mut y_edges = vec![
-			selection_rect.min.y,
-			selection_rect.min.y + interior_reach_y,
-			selection_rect.center().y,
-			selection_rect.max.y - interior_reach_y,
-			selection_rect.max.y,
-		];
-
-		for handle in WindowRenderer::frozen_selection_resize_handles(capture_rect) {
-			x_edges.push(handle.hit_rect.min.x);
-			x_edges.push(handle.hit_rect.max.x);
-			y_edges.push(handle.hit_rect.min.y);
-			y_edges.push(handle.hit_rect.max.y);
-		}
-
-		sort_unique_axis_values(&mut x_edges);
-		sort_unique_axis_values(&mut y_edges);
-
-		let mut rects = Vec::new();
-
-		for x_pair in x_edges.windows(2) {
-			let [min_x, max_x] = [x_pair[0], x_pair[1]];
-
-			if max_x <= min_x {
-				continue;
-			}
-
-			for y_pair in y_edges.windows(2) {
-				let [min_y, max_y] = [y_pair[0], y_pair[1]];
-
-				if max_y <= min_y {
-					continue;
-				}
-
-				let rect = Rect::from_min_max(Pos2::new(min_x, min_y), Pos2::new(max_x, max_y));
-				let point = rect.center();
-				let Some(interaction) = Self::frozen_selection_interaction_kind(
-					capture_rect,
-					point.x as u32,
-					point.y as u32,
-				) else {
-					continue;
-				};
-				let FrozenSelectionInteractionKind::Resize(corner) = interaction else {
-					rects.push(OverlayCursorRect::new(rect, CursorIcon::Grab));
-
-					continue;
-				};
-				let icon = Self::frozen_selection_resize_cursor_icon(corner);
-				let mut adjusted_min = rect.min;
-				let mut adjusted_max = rect.max;
-
-				if WindowRenderer::frozen_selection_resize_hit_test(
-					capture_rect,
-					Pos2::new(rect.min.x, point.y),
-				) != Some(corner)
-				{
-					adjusted_min.x = trim_rect_min_edge(adjusted_min.x, adjusted_max.x);
-				}
-				if WindowRenderer::frozen_selection_resize_hit_test(
-					capture_rect,
-					Pos2::new(rect.max.x, point.y),
-				) != Some(corner)
-				{
-					adjusted_max.x = trim_rect_max_edge(adjusted_max.x, adjusted_min.x);
-				}
-				if WindowRenderer::frozen_selection_resize_hit_test(
-					capture_rect,
-					Pos2::new(point.x, rect.min.y),
-				) != Some(corner)
-				{
-					adjusted_min.y = trim_rect_min_edge(adjusted_min.y, adjusted_max.y);
-				}
-				if WindowRenderer::frozen_selection_resize_hit_test(
-					capture_rect,
-					Pos2::new(point.x, rect.max.y),
-				) != Some(corner)
-				{
-					adjusted_max.y = trim_rect_max_edge(adjusted_max.y, adjusted_min.y);
-				}
-				if adjusted_max.x <= adjusted_min.x || adjusted_max.y <= adjusted_min.y {
-					continue;
-				}
-
-				rects.push(OverlayCursorRect::new(
-					Rect::from_min_max(adjusted_min, adjusted_max),
-					icon,
-				));
-			}
-		}
-
-		rects
-	}
-
-	fn sync_overlay_cursor_icons(&self) {
-		for overlay_window in self.windows.values() {
-			#[cfg(not(target_os = "macos"))]
-			let icon = self.overlay_cursor_icon_for_monitor(overlay_window.monitor);
-
-			#[cfg(not(target_os = "macos"))]
-			overlay_window.window.set_cursor(icon);
-			#[cfg(target_os = "macos")]
-			overlay_window.cursor_rects.sync_cursor_rects(
-				overlay_window.window.as_ref(),
-				&self.frozen_selection_cursor_rects_for_monitor(overlay_window.monitor),
-			);
-		}
-	}
-
-	pub(super) fn frozen_selection_drag_hides_auxiliary_windows(&self) -> bool {
-		matches!(self.state.mode, OverlayMode::Frozen) && self.frozen_selection_drag.active
-	}
-
-	pub(super) fn live_drag_hides_auxiliary_windows(&self) -> bool {
-		matches!(self.state.mode, OverlayMode::Live) && self.state.drag_rect.is_some()
-	}
-
-	fn hide_auxiliary_windows_for_frozen_selection_drag(&mut self) {
-		if let Some(hud_window) = self.hud_window.as_ref() {
-			hud_window.window.set_visible(false);
-		}
-
-		self.hud_window_visible = false;
-
-		if let Some(loupe_window) = self.loupe_window.as_ref() {
-			loupe_window.window.set_visible(false);
-		}
-
-		self.loupe_window_visible = false;
-
-		self.reset_loupe_window_warmup_redraws();
-
-		if let Some(toolbar_window) = self.toolbar_window.as_ref() {
-			toolbar_window.window.set_visible(false);
-		}
-
-		self.skip_toolbar_focus_on_next_show = true;
-		self.toolbar_window_visible = false;
-		self.toolbar_window_warmup_redraws_remaining = 0;
-
-		if let Some(preview_window) = self.scroll_preview_window.as_ref() {
-			preview_window.window.set_visible(false);
-		}
-
-		self.last_present_at = Instant::now();
-	}
-
-	fn hide_auxiliary_windows_for_live_drag(&mut self) {
-		if let Some(hud_window) = self.hud_window.as_ref() {
-			hud_window.window.set_visible(false);
-		}
-
-		self.hud_window_visible = false;
-
-		if let Some(loupe_window) = self.loupe_window.as_ref() {
-			loupe_window.window.set_visible(false);
-		}
-
-		self.loupe_window_visible = false;
-
-		self.reset_loupe_window_warmup_redraws();
-
-		self.last_present_at = Instant::now();
-	}
-
-	fn stop_frozen_selection_drag(&mut self) {
-		let was_active = self.frozen_selection_drag.active;
-
-		self.frozen_selection_drag = FrozenSelectionDragState::default();
-
-		if was_active && let Some(monitor) = self.state.monitor {
-			self.request_redraw_for_monitor(monitor);
-		}
-	}
-
-	fn stop_frozen_mosaic_drag(&mut self) {
-		self.frozen_mosaic_drag = FrozenMosaicDragState::default();
-		self.state.frozen_mosaic_preview_rect = None;
-	}
-
-	fn begin_frozen_brush_stroke(&mut self, global: GlobalPoint) -> bool {
-		if self.toolbar_state.selected_tool != FrozenToolbarTool::Pen {
-			return false;
-		}
-		if self.frozen_brush.active_stroke.is_some() {
-			return false;
-		}
-
-		let Some((monitor, capture_rect)) = self.frozen_brush_capture_target() else {
-			return false;
-		};
-		let Some((cursor_x, cursor_y)) = monitor.local_u32(global) else {
-			return false;
-		};
-
-		if !capture_rect.contains((cursor_x, cursor_y)) {
-			return false;
-		}
-
-		let point = Pos2::new(cursor_x as f32, cursor_y as f32);
-		let sampled_at = Instant::now();
-
-		self.frozen_brush.active_stroke =
-			Some(Self::new_active_frozen_brush_stroke(point, sampled_at));
-
-		self.request_redraw_for_monitor(monitor);
-
-		true
-	}
-
-	fn update_frozen_brush_stroke(&mut self, global: GlobalPoint) -> bool {
-		let Some((monitor, capture_rect)) = self.frozen_brush_capture_target() else {
-			return false;
-		};
-		let Some((cursor_x, cursor_y)) = monitor.local_u32(global) else {
-			return false;
-		};
-		let Some(active_stroke) = self.frozen_brush.active_stroke.as_mut() else {
-			return false;
-		};
-		let point = Self::clamped_point_in_capture_rect(capture_rect, cursor_x, cursor_y);
-		let sampled_at = Instant::now();
-
-		if let Some(previous) = active_stroke.raw_points.last().copied()
-			&& previous.distance(point) < FROZEN_BRUSH_POINT_SPACING_MIN_POINTS
-		{
-			return false;
-		}
-
-		Self::append_frozen_brush_raw_sample(active_stroke, point, sampled_at);
-
-		true
-	}
-
-	fn finish_frozen_brush_stroke(&mut self) -> bool {
-		let Some(stroke) = self.frozen_brush.active_stroke.take() else {
-			return false;
-		};
-
-		if stroke.points.is_empty() {
-			return false;
-		}
-
-		self.frozen_brush
-			.committed_strokes
-			.push(FrozenBrushStroke { points: Self::finished_frozen_brush_points(&stroke) });
-		self.push_frozen_edit_to_undo_history(FrozenEditKind::BrushStroke);
-		self.sync_frozen_toolbar_state();
-
-		if let Some(monitor) = self.state.monitor {
-			self.request_redraw_for_monitor(monitor);
-		}
-
-		true
-	}
-
-	fn undo_frozen_brush_stroke(&mut self) -> bool {
-		let Some(stroke) = self.frozen_brush.committed_strokes.pop() else {
-			return false;
-		};
-
-		self.frozen_brush.redo_strokes.push(stroke);
-		self.sync_frozen_toolbar_state();
-
-		self.toolbar_state.needs_redraw = true;
-
-		self.request_redraw_toolbar_window();
-
-		if let Some(monitor) = self.state.monitor {
-			self.request_redraw_for_monitor(monitor);
-		}
-
-		true
-	}
-
-	fn redo_frozen_brush_stroke(&mut self) -> bool {
-		let Some(stroke) = self.frozen_brush.redo_strokes.pop() else {
-			return false;
-		};
-
-		self.frozen_brush.committed_strokes.push(stroke);
-		self.sync_frozen_toolbar_state();
-
-		self.toolbar_state.needs_redraw = true;
-
-		self.request_redraw_toolbar_window();
-
-		if let Some(monitor) = self.state.monitor {
-			self.request_redraw_for_monitor(monitor);
-		}
-
-		true
-	}
-
-	fn clamped_point_in_capture_rect(
-		capture_rect: RectPoints,
-		cursor_x: u32,
-		cursor_y: u32,
-	) -> Pos2 {
-		let max_x = capture_rect.x.saturating_add(capture_rect.width.saturating_sub(1));
-		let max_y = capture_rect.y.saturating_add(capture_rect.height.saturating_sub(1));
-
-		Pos2::new(
-			cursor_x.clamp(capture_rect.x, max_x) as f32,
-			cursor_y.clamp(capture_rect.y, max_y) as f32,
-		)
-	}
-
-	fn new_active_frozen_brush_stroke(point: Pos2, sampled_at: Instant) -> ActiveFrozenBrushStroke {
-		ActiveFrozenBrushStroke {
-			raw_points: vec![point],
-			points: vec![point],
-			model_state: FrozenBrushModelState {
-				filtered_input_point: point,
-				modeled_point: point,
-				modeled_velocity: Vec2::ZERO,
-				modeled_elapsed_seconds: 0.0,
-			},
-			started_at: sampled_at,
-			last_sample_at: sampled_at,
-		}
-	}
-
-	fn append_frozen_brush_raw_sample(
-		active_stroke: &mut ActiveFrozenBrushStroke,
-		point: Pos2,
-		sampled_at: Instant,
-	) {
-		let delta_seconds = sampled_at
-			.saturating_duration_since(active_stroke.last_sample_at)
-			.as_secs_f32()
-			.max(1.0 / 240.0);
-
-		active_stroke.raw_points.push(point);
-
-		let snap_boost = Self::frozen_brush_feature_snap_boost(&active_stroke.raw_points);
-		let response = Self::frozen_brush_input_response_with_feature_boost(
-			&active_stroke.raw_points,
-			delta_seconds,
-			Self::frozen_brush_response_floor_boost(&active_stroke.raw_points, snap_boost),
-		);
-
-		active_stroke.model_state.filtered_input_point = Pos2::new(
-			active_stroke.model_state.filtered_input_point.x
-				+ ((point.x - active_stroke.model_state.filtered_input_point.x) * response),
-			active_stroke.model_state.filtered_input_point.y
-				+ ((point.y - active_stroke.model_state.filtered_input_point.y) * response),
-		);
-
-		let elapsed_seconds =
-			sampled_at.saturating_duration_since(active_stroke.started_at).as_secs_f32();
-
-		Self::advance_frozen_brush_model(
-			&mut active_stroke.points,
-			&mut active_stroke.model_state,
-			elapsed_seconds,
-		);
-
-		if snap_boost >= 1.0 {
-			active_stroke.model_state.modeled_point = Pos2::new(
-				active_stroke.model_state.modeled_point.x
-					+ (point.x - active_stroke.model_state.modeled_point.x),
-				active_stroke.model_state.modeled_point.y
-					+ (point.y - active_stroke.model_state.modeled_point.y),
-			);
-			active_stroke.model_state.modeled_velocity *= 0.35;
-
-			Self::push_modeled_frozen_brush_point(
-				&mut active_stroke.points,
-				active_stroke.model_state.modeled_point,
-			);
-		} else if snap_boost > 0.0 {
-			active_stroke.model_state.modeled_point = Pos2::new(
-				active_stroke.model_state.modeled_point.x
-					+ ((point.x - active_stroke.model_state.modeled_point.x) * 0.24),
-				active_stroke.model_state.modeled_point.y
-					+ ((point.y - active_stroke.model_state.modeled_point.y) * 0.24),
-			);
-			active_stroke.model_state.modeled_velocity *= 0.82;
-
-			Self::push_modeled_frozen_brush_point(
-				&mut active_stroke.points,
-				active_stroke.model_state.modeled_point,
-			);
-		}
-
-		active_stroke.last_sample_at = sampled_at;
-	}
-
-	#[cfg(test)]
-	fn frozen_brush_input_response(points: &[Pos2], delta_seconds: f32) -> f32 {
-		let snap_boost = Self::frozen_brush_feature_snap_boost(points);
-
-		Self::frozen_brush_input_response_with_feature_boost(
-			points,
-			delta_seconds,
-			Self::frozen_brush_response_floor_boost(points, snap_boost),
-		)
-	}
-
-	fn frozen_brush_input_response_with_feature_boost(
-		points: &[Pos2],
-		delta_seconds: f32,
-		feature_boost: f32,
-	) -> f32 {
-		let speed_points_per_second = points
-			.windows(2)
-			.last()
-			.map_or(0.0, |window| window[0].distance(window[1]) / delta_seconds);
-		let normalized_speed = ((speed_points_per_second
-			- FROZEN_BRUSH_MODEL_SPEED_FLOOR_POINTS_PER_SECOND)
-			/ (FROZEN_BRUSH_MODEL_SPEED_CEILING_POINTS_PER_SECOND
-				- FROZEN_BRUSH_MODEL_SPEED_FLOOR_POINTS_PER_SECOND))
-			.clamp(0.0, 1.0);
-		let base_response = FROZEN_BRUSH_MODEL_INPUT_RESPONSE_MIN
-			+ ((FROZEN_BRUSH_MODEL_INPUT_RESPONSE_MAX - FROZEN_BRUSH_MODEL_INPUT_RESPONSE_MIN)
-				* normalized_speed);
-
-		base_response.max(feature_boost)
-	}
-
-	fn frozen_brush_response_floor_boost(points: &[Pos2], snap_boost: f32) -> f32 {
-		Self::frozen_brush_curve_response_boost(points).max(snap_boost)
-	}
-
-	fn frozen_brush_preview_rounding_passes(points: &[Pos2]) -> usize {
-		if Self::frozen_brush_has_sustained_curve_context(points) {
-			FROZEN_BRUSH_PREVIEW_ROUNDING_PASSES + 1
-		} else {
-			FROZEN_BRUSH_PREVIEW_ROUNDING_PASSES
-		}
-	}
-
-	fn frozen_brush_curve_response_boost(points: &[Pos2]) -> f32 {
-		if Self::frozen_brush_has_sustained_curve_context(points) {
-			FROZEN_BRUSH_MODEL_CURVE_RESPONSE_BOOST
-		} else {
-			0.0
-		}
-	}
-
-	fn frozen_brush_feature_snap_boost(points: &[Pos2]) -> f32 {
-		let len = points.len();
-
-		if len < 3 {
-			return 0.0;
-		}
-
-		let previous = points[len - 2];
-		let current = points[len - 1];
-		let anchor = points[len - 3];
-		let current_turn_angle = Self::frozen_brush_turn_angle(anchor, previous, current);
-		let current_amplitude =
-			Self::frozen_brush_point_to_segment_distance(previous, anchor, current);
-
-		if current_turn_angle < FROZEN_BRUSH_MODEL_FEATURE_TURN_RADIANS
-			|| current_amplitude < FROZEN_BRUSH_MODEL_FEATURE_AMPLITUDE_POINTS
-		{
-			return 0.0;
-		}
-		if Self::frozen_brush_has_sustained_curve_context(points) {
-			return 0.0;
-		}
-		if current_turn_angle >= FROZEN_BRUSH_MODEL_SHARP_TURN_RADIANS {
-			return 1.0;
-		}
-		if len < 4 {
-			return 0.0;
-		}
-
-		let support = points[len - 4];
-		let previous_turn = Self::frozen_brush_signed_turn(support, anchor, previous);
-		let current_turn = Self::frozen_brush_signed_turn(anchor, previous, current);
-		let previous_turn_angle = Self::frozen_brush_turn_angle(support, anchor, previous);
-		let previous_amplitude =
-			Self::frozen_brush_point_to_segment_distance(anchor, support, previous);
-
-		if previous_turn * current_turn < 0.0
-			&& previous_turn_angle >= FROZEN_BRUSH_MODEL_FEATURE_TURN_RADIANS
-			&& previous_amplitude >= FROZEN_BRUSH_MODEL_FEATURE_AMPLITUDE_POINTS
-		{
-			return 0.94;
-		}
-
-		0.0
-	}
-
-	fn frozen_brush_has_sustained_curve_context(points: &[Pos2]) -> bool {
-		let len = points.len();
-
-		if len < 5 {
-			return false;
-		}
-
-		let recent_points = &points[len - 5..];
-		let mut turn_sign: f32 = 0.0;
-
-		for window in recent_points.windows(3) {
-			let signed_turn = Self::frozen_brush_signed_turn(window[0], window[1], window[2]);
-			let turn_angle = Self::frozen_brush_turn_angle(window[0], window[1], window[2]);
-			let amplitude =
-				Self::frozen_brush_point_to_segment_distance(window[1], window[0], window[2]);
-
-			if turn_angle < FROZEN_BRUSH_MODEL_CURVE_TURN_RADIANS
-				|| amplitude < FROZEN_BRUSH_MODEL_CURVE_AMPLITUDE_POINTS
-			{
-				return false;
-			}
-			if turn_sign.abs() <= f32::EPSILON {
-				turn_sign = signed_turn.signum();
-			} else if signed_turn.signum() != turn_sign {
-				return false;
-			}
-		}
-
-		true
-	}
-
-	fn advance_frozen_brush_model(
-		points: &mut Vec<Pos2>,
-		model_state: &mut FrozenBrushModelState,
-		target_elapsed_seconds: f32,
-	) {
-		if target_elapsed_seconds <= model_state.modeled_elapsed_seconds {
-			return;
-		}
-
-		while model_state.modeled_elapsed_seconds + FROZEN_BRUSH_MODEL_TIMESTEP_SECONDS
-			< target_elapsed_seconds
-		{
-			Self::step_frozen_brush_model(points, model_state, FROZEN_BRUSH_MODEL_TIMESTEP_SECONDS);
-		}
-
-		let remainder = target_elapsed_seconds - model_state.modeled_elapsed_seconds;
-
-		if remainder > f32::EPSILON {
-			Self::step_frozen_brush_model(points, model_state, remainder);
-		}
-	}
-
-	fn step_frozen_brush_model(
-		points: &mut Vec<Pos2>,
-		model_state: &mut FrozenBrushModelState,
-		delta_seconds: f32,
-	) {
-		let displacement = model_state.filtered_input_point - model_state.modeled_point;
-		let acceleration = (displacement * FROZEN_BRUSH_MODEL_SPRING_CONSTANT)
-			- (model_state.modeled_velocity * FROZEN_BRUSH_MODEL_DRAG_CONSTANT);
-
-		model_state.modeled_velocity += acceleration * delta_seconds;
-		model_state.modeled_point += model_state.modeled_velocity * delta_seconds;
-		model_state.modeled_elapsed_seconds += delta_seconds;
-
-		Self::push_modeled_frozen_brush_point(points, model_state.modeled_point);
-	}
-
-	fn finished_frozen_brush_points(stroke: &ActiveFrozenBrushStroke) -> Vec<Pos2> {
-		let source_points =
-			if stroke.points.len() >= 2 { &stroke.points } else { &stroke.raw_points };
-
-		Self::processed_frozen_brush_points(
-			source_points,
-			stroke.raw_points[0],
-			stroke.raw_points[stroke.raw_points.len().saturating_sub(1)],
-			FROZEN_BRUSH_COMMIT_ROUNDING_PASSES,
-			true,
-		)
-	}
-
-	fn rendered_frozen_brush_points(points: &[Pos2], sample_step: f32) -> Vec<Pos2> {
-		match points {
-			[] => Vec::new(),
-			[first] => vec![*first],
-			[first, second] => {
-				let sample_step = sample_step.max(0.1);
-				let mut rendered = vec![*first];
-
-				Self::append_frozen_brush_linear_segment(
-					&mut rendered,
-					*first,
-					*second,
-					sample_step,
-				);
-
-				rendered
-			},
-			_ => {
-				let sample_step = sample_step.max(0.1);
-				let mut rendered = Vec::with_capacity(points.len() * 6);
-
-				rendered.push(points[0]);
-
-				for index in 0..points.len().saturating_sub(1) {
-					let previous = if index == 0 { points[0] } else { points[index - 1] };
-					let start = points[index];
-					let end = points[index + 1];
-					let next = points
-						.get(index + 2)
-						.copied()
-						.unwrap_or(points[points.len().saturating_sub(1)]);
-
-					Self::append_frozen_brush_curve_segment(
-						&mut rendered,
-						previous,
-						start,
-						end,
-						next,
-						sample_step,
-					);
-				}
-
-				rendered
-			},
-		}
-	}
-
-	fn active_frozen_brush_display_points(active_stroke: &ActiveFrozenBrushStroke) -> Vec<Pos2> {
-		let source_points = if active_stroke.points.len() >= 2 {
-			&active_stroke.points
-		} else {
-			&active_stroke.raw_points
-		};
-		let rounding_passes = Self::frozen_brush_preview_rounding_passes(&active_stroke.raw_points);
-
-		Self::processed_frozen_brush_points(
-			source_points,
-			active_stroke.raw_points[0],
-			active_stroke.raw_points[active_stroke.raw_points.len().saturating_sub(1)],
-			rounding_passes,
-			false,
-		)
-	}
-
-	fn preview_frozen_brush_points(active_stroke: &ActiveFrozenBrushStroke) -> Vec<Pos2> {
-		Self::active_frozen_brush_display_points(active_stroke)
-	}
-
-	fn processed_frozen_brush_points(
-		source_points: &[Pos2],
-		start_point: Pos2,
-		end_point: Pos2,
-		rounding_passes: usize,
-		streamline: bool,
-	) -> Vec<Pos2> {
-		match source_points {
-			[] => Vec::new(),
-			[_] | [_, _] => vec![start_point, end_point],
-			_ => {
-				let streamlined = if streamline {
-					Self::streamlined_frozen_brush_points(source_points)
-				} else {
-					source_points.to_vec()
-				};
-				let mut rounded =
-					Self::rounded_open_frozen_brush_points(&streamlined, rounding_passes);
-
-				if rounded.len() < 2 {
-					return vec![start_point, end_point];
-				}
-
-				rounded[0] = start_point;
-
-				if let Some(last) = rounded.last_mut() {
-					*last = end_point;
-				}
-
-				rounded
-			},
-		}
-	}
-
-	fn streamlined_frozen_brush_points(raw_points: &[Pos2]) -> Vec<Pos2> {
-		let Some(first) = raw_points.first().copied() else {
-			return Vec::new();
-		};
-		let mut streamlined = vec![first];
-		let mut filtered = first;
-
-		for window in raw_points.windows(2) {
-			let previous_raw = window[0];
-			let current_raw = window[1];
-			let normalized_distance = ((previous_raw.distance(current_raw)
-				- FROZEN_BRUSH_POINT_SPACING_MIN_POINTS)
-				/ (FROZEN_BRUSH_STREAMLINE_DISTANCE_CEILING_POINTS
-					- FROZEN_BRUSH_POINT_SPACING_MIN_POINTS))
-				.clamp(0.0, 1.0);
-			let response = FROZEN_BRUSH_STREAMLINE_RESPONSE_MIN
-				+ ((FROZEN_BRUSH_STREAMLINE_RESPONSE_MAX - FROZEN_BRUSH_STREAMLINE_RESPONSE_MIN)
-					* normalized_distance);
-
-			filtered = Pos2::new(
-				filtered.x + ((current_raw.x - filtered.x) * response),
-				filtered.y + ((current_raw.y - filtered.y) * response),
-			);
-
-			Self::push_processed_frozen_brush_point(&mut streamlined, filtered);
-		}
-
-		streamlined[0] = raw_points[0];
-
-		let last_raw = raw_points[raw_points.len().saturating_sub(1)];
-
-		if let Some(last) = streamlined.last_mut() {
-			*last = last_raw;
-		}
-
-		streamlined
-	}
-
-	fn rounded_open_frozen_brush_points(points: &[Pos2], passes: usize) -> Vec<Pos2> {
-		if points.len() <= 2 || passes == 0 {
-			return points.to_vec();
-		}
-
-		let mut rounded = points.to_vec();
-
-		for _ in 0..passes {
-			if rounded.len() <= 2 {
-				break;
-			}
-
-			let mut next = Vec::with_capacity((rounded.len() * 2).saturating_sub(2));
-
-			next.push(rounded[0]);
-
-			for window in rounded.windows(2) {
-				let start = window[0];
-				let end = window[1];
-				let quarter =
-					Pos2::new((start.x * 0.75) + (end.x * 0.25), (start.y * 0.75) + (end.y * 0.25));
-				let three_quarters =
-					Pos2::new((start.x * 0.25) + (end.x * 0.75), (start.y * 0.25) + (end.y * 0.75));
-
-				Self::push_processed_frozen_brush_point(&mut next, quarter);
-				Self::push_processed_frozen_brush_point(&mut next, three_quarters);
-			}
-
-			let last = rounded[rounded.len().saturating_sub(1)];
-
-			if next.last().is_none_or(|point| {
-				point.distance(last) > FROZEN_BRUSH_PREVIEW_POINT_SPACING_MIN_POINTS
-			}) {
-				next.push(last);
-			} else if let Some(last_point) = next.last_mut() {
-				*last_point = last;
-			}
-
-			rounded = next;
-		}
-
-		rounded
-	}
-
-	fn append_frozen_brush_curve_segment(
-		points: &mut Vec<Pos2>,
-		previous: Pos2,
-		start: Pos2,
-		end: Pos2,
-		next: Pos2,
-		sample_step: f32,
-	) {
-		let approximate_length = start.distance(end);
-		let steps = ((approximate_length / sample_step).ceil().max(1.0)) as usize;
-
-		for step in 1..=steps {
-			let t = step as f32 / steps as f32;
-			let point = Self::catmull_rom_frozen_brush_point(previous, start, end, next, t);
-
-			Self::push_frozen_brush_sample_point(points, point);
-		}
-	}
-
-	fn append_frozen_brush_linear_segment(
-		points: &mut Vec<Pos2>,
-		start: Pos2,
-		end: Pos2,
-		sample_step: f32,
-	) {
-		let approximate_length = start.distance(end);
-		let steps = ((approximate_length / sample_step).ceil().max(1.0)) as usize;
-
-		for step in 1..=steps {
-			let t = step as f32 / steps as f32;
-			let point = Pos2::new(start.x + (end.x - start.x) * t, start.y + (end.y - start.y) * t);
-
-			Self::push_frozen_brush_sample_point(points, point);
-		}
-	}
-
-	fn catmull_rom_frozen_brush_point(
-		previous: Pos2,
-		start: Pos2,
-		end: Pos2,
-		next: Pos2,
-		t: f32,
-	) -> Pos2 {
-		const CENTRIPETAL_ALPHA: f32 = 0.5;
-		const MIN_PARAMETER_STEP: f32 = 1.0e-3;
-
-		let t0 = 0.0;
-		let t1 = t0 + previous.distance(start).powf(CENTRIPETAL_ALPHA).max(MIN_PARAMETER_STEP);
-		let t2 = t1 + start.distance(end).powf(CENTRIPETAL_ALPHA).max(MIN_PARAMETER_STEP);
-		let t3 = t2 + end.distance(next).powf(CENTRIPETAL_ALPHA).max(MIN_PARAMETER_STEP);
-		let sample_t = t1 + ((t2 - t1) * t.clamp(0.0, 1.0));
-		let a1 = Self::centripetal_catmull_rom_lerp(previous, start, t0, t1, sample_t);
-		let a2 = Self::centripetal_catmull_rom_lerp(start, end, t1, t2, sample_t);
-		let a3 = Self::centripetal_catmull_rom_lerp(end, next, t2, t3, sample_t);
-		let b1 = Self::centripetal_catmull_rom_lerp(a1, a2, t0, t2, sample_t);
-		let b2 = Self::centripetal_catmull_rom_lerp(a2, a3, t1, t3, sample_t);
-
-		Self::centripetal_catmull_rom_lerp(b1, b2, t1, t2, sample_t)
-	}
-
-	fn centripetal_catmull_rom_lerp(
-		start: Pos2,
-		end: Pos2,
-		start_t: f32,
-		end_t: f32,
-		sample_t: f32,
-	) -> Pos2 {
-		let span = (end_t - start_t).max(f32::EPSILON);
-		let start_weight = (end_t - sample_t) / span;
-		let end_weight = (sample_t - start_t) / span;
-
-		Pos2::new(
-			(start.x * start_weight) + (end.x * end_weight),
-			(start.y * start_weight) + (end.y * end_weight),
-		)
-	}
-
-	fn push_frozen_brush_sample_point(points: &mut Vec<Pos2>, point: Pos2) {
-		if points.last().is_none_or(|previous| previous.distance(point) > f32::EPSILON) {
-			points.push(point);
-		}
-	}
-
-	fn push_modeled_frozen_brush_point(points: &mut Vec<Pos2>, point: Pos2) {
-		let Some(previous) = points.last().copied() else {
-			points.push(point);
-
-			return;
-		};
-
-		if previous.distance(point) < f32::EPSILON {
-			return;
-		}
-		if previous.distance(point) < FROZEN_BRUSH_MODELED_POINT_SPACING_MIN_POINTS
-			&& points.len() > 1
-		{
-			if let Some(last) = points.last_mut() {
-				*last = point;
-			}
-
-			return;
-		}
-
-		points.push(point);
-	}
-
-	fn push_processed_frozen_brush_point(points: &mut Vec<Pos2>, point: Pos2) {
-		let Some(previous) = points.last().copied() else {
-			points.push(point);
-
-			return;
-		};
-
-		if previous.distance(point) < FROZEN_BRUSH_PREVIEW_POINT_SPACING_MIN_POINTS {
-			if let Some(last) = points.last_mut() {
-				*last = point;
-			}
-
-			return;
-		}
-
-		points.push(point);
-	}
-
-	#[cfg(test)]
-	fn corrected_frozen_brush_points(points: &[Pos2]) -> Vec<Pos2> {
-		if points.len() <= 2 {
-			return points.to_vec();
-		}
-
-		let started_at = Instant::now();
-		let mut stroke = Self::new_active_frozen_brush_stroke(points[0], started_at);
-
-		for (index, point) in points.iter().copied().enumerate().skip(1) {
-			let sampled_at = started_at
-				+ Duration::from_secs_f32(
-					index as f32 * FROZEN_BRUSH_MODEL_SYNTHETIC_SAMPLE_INTERVAL_SECONDS,
-				);
-
-			Self::append_frozen_brush_raw_sample(&mut stroke, point, sampled_at);
-		}
-
-		Self::finished_frozen_brush_points(&stroke)
-	}
-
-	fn frozen_brush_point_to_segment_distance(point: Pos2, start: Pos2, end: Pos2) -> f32 {
-		let segment_x = end.x - start.x;
-		let segment_y = end.y - start.y;
-		let segment_length_sq = (segment_x * segment_x) + (segment_y * segment_y);
-
-		if segment_length_sq <= f32::EPSILON {
-			return point.distance(start);
-		}
-
-		let t = (((point.x - start.x) * segment_x) + ((point.y - start.y) * segment_y))
-			/ segment_length_sq;
-		let t = t.clamp(0.0, 1.0);
-		let projection = Pos2::new(start.x + (segment_x * t), start.y + (segment_y * t));
-
-		point.distance(projection)
-	}
-
-	fn frozen_brush_turn_angle(previous: Pos2, current: Pos2, next: Pos2) -> f32 {
-		let first = current - previous;
-		let second = next - current;
-		let first_length = first.length();
-		let second_length = second.length();
-
-		if first_length <= f32::EPSILON || second_length <= f32::EPSILON {
-			return 0.0;
-		}
-
-		let cosine = (first.dot(second) / (first_length * second_length)).clamp(-1.0, 1.0);
-
-		cosine.acos()
-	}
-
-	fn frozen_brush_signed_turn(previous: Pos2, current: Pos2, next: Pos2) -> f32 {
-		let first = current - previous;
-		let second = next - current;
-
-		(first.x * second.y) - (first.y * second.x)
-	}
-
-	fn update_frozen_selection_drag_rect(&mut self, global: GlobalPoint) -> bool {
-		if !self.frozen_selection_drag.active {
-			return false;
-		}
-
-		let Some((monitor, _capture_rect)) = self.frozen_selection_drag_target() else {
-			self.stop_frozen_selection_drag();
-
-			return false;
-		};
-		let anchor_rect = self.frozen_selection_drag.anchor_rect;
-		let next_rect = match self.frozen_selection_drag.interaction {
-			FrozenSelectionInteractionKind::Move => {
-				let (cursor_x, cursor_y) = Self::clamped_local_point_in_monitor(monitor, global);
-				let desired_x =
-					i64::from(cursor_x) - i64::from(self.frozen_selection_drag.pointer_offset_x);
-				let desired_y =
-					i64::from(cursor_y) - i64::from(self.frozen_selection_drag.pointer_offset_y);
-
-				Self::clamp_frozen_capture_rect_to_monitor(
-					monitor,
-					anchor_rect.width,
-					anchor_rect.height,
-					desired_x,
-					desired_y,
-				)
-			},
-			FrozenSelectionInteractionKind::Resize(corner) => {
-				let (cursor_x, cursor_y) = Self::local_point_in_monitor_space(monitor, global);
-
-				Self::resize_frozen_capture_rect_from_corner(
-					monitor,
-					anchor_rect,
-					corner,
-					self.frozen_selection_drag.press_cursor_x,
-					self.frozen_selection_drag.press_cursor_y,
-					cursor_x,
-					cursor_y,
-				)
-			},
-		};
-
-		self.apply_frozen_capture_rect_update(monitor, next_rect)
-	}
-
-	fn update_frozen_mosaic_drag_rect(&mut self, global: GlobalPoint) -> bool {
-		if !self.frozen_mosaic_drag.active {
-			return false;
-		}
-
-		let Some((monitor, capture_rect)) = self.frozen_mosaic_drag_target() else {
-			self.stop_frozen_mosaic_drag();
-
-			return false;
-		};
-		let (cursor_x, cursor_y) = Self::clamped_local_point_in_rect(monitor, capture_rect, global);
-		let next_rect = Self::rect_from_drag_points(
-			self.frozen_mosaic_drag.anchor_x,
-			self.frozen_mosaic_drag.anchor_y,
-			cursor_x,
-			cursor_y,
-		);
-
-		if self.state.frozen_mosaic_preview_rect == Some(next_rect) {
-			return false;
-		}
-
-		self.state.frozen_mosaic_preview_rect = Some(next_rect);
-
-		self.request_redraw_for_monitor(monitor);
-
-		true
-	}
-
-	fn clamped_local_point_in_monitor(monitor: MonitorRect, global: GlobalPoint) -> (u32, u32) {
-		let max_x = i64::from(monitor.width.saturating_sub(1));
-		let max_y = i64::from(monitor.height.saturating_sub(1));
-		let local_x = (i64::from(global.x) - i64::from(monitor.origin.x)).clamp(0, max_x) as u32;
-		let local_y = (i64::from(global.y) - i64::from(monitor.origin.y)).clamp(0, max_y) as u32;
-
-		(local_x, local_y)
-	}
-
-	fn clamped_local_point_in_rect(
-		monitor: MonitorRect,
-		capture_rect: RectPoints,
-		global: GlobalPoint,
-	) -> (u32, u32) {
-		let (local_x, local_y) = Self::clamped_local_point_in_monitor(monitor, global);
-		let max_x = capture_rect.x.saturating_add(capture_rect.width.saturating_sub(1));
-		let max_y = capture_rect.y.saturating_add(capture_rect.height.saturating_sub(1));
-
-		(local_x.clamp(capture_rect.x, max_x), local_y.clamp(capture_rect.y, max_y))
-	}
-
-	fn rect_from_drag_points(
-		anchor_x: u32,
-		anchor_y: u32,
-		cursor_x: u32,
-		cursor_y: u32,
-	) -> RectPoints {
-		let left = anchor_x.min(cursor_x);
-		let top = anchor_y.min(cursor_y);
-		let right = anchor_x.max(cursor_x).saturating_add(1);
-		let bottom = anchor_y.max(cursor_y).saturating_add(1);
-
-		RectPoints::new(left, top, right.saturating_sub(left), bottom.saturating_sub(top))
-	}
-
-	fn local_point_in_monitor_space(monitor: MonitorRect, global: GlobalPoint) -> (i64, i64) {
-		(
-			i64::from(global.x) - i64::from(monitor.origin.x),
-			i64::from(global.y) - i64::from(monitor.origin.y),
-		)
-	}
-
-	fn clamp_frozen_capture_rect_to_monitor(
-		monitor: MonitorRect,
-		width: u32,
-		height: u32,
-		desired_x: i64,
-		desired_y: i64,
-	) -> RectPoints {
-		let max_x = i64::from(monitor.width.saturating_sub(width));
-		let max_y = i64::from(monitor.height.saturating_sub(height));
-		let x = desired_x.clamp(0, max_x) as u32;
-		let y = desired_y.clamp(0, max_y) as u32;
-
-		RectPoints::new(x, y, width, height)
-	}
-
-	fn resize_frozen_capture_rect_from_corner(
-		monitor: MonitorRect,
-		anchor_rect: RectPoints,
-		corner: FrozenSelectionCorner,
-		press_cursor_x: u32,
-		press_cursor_y: u32,
-		cursor_x: i64,
-		cursor_y: i64,
-	) -> RectPoints {
-		let left = i64::from(anchor_rect.x);
-		let top = i64::from(anchor_rect.y);
-		let right = i64::from(anchor_rect.x.saturating_add(anchor_rect.width));
-		let bottom = i64::from(anchor_rect.y.saturating_add(anchor_rect.height));
-		let delta_x = cursor_x - i64::from(press_cursor_x);
-		let delta_y = cursor_y - i64::from(press_cursor_y);
-		let monitor_width = i64::from(monitor.width);
-		let monitor_height = i64::from(monitor.height);
-
-		match corner {
-			FrozenSelectionCorner::TopLeft => {
-				let next_left = (left + delta_x).clamp(0, right.saturating_sub(1)) as u32;
-				let next_top = (top + delta_y).clamp(0, bottom.saturating_sub(1)) as u32;
-
-				RectPoints::new(
-					next_left,
-					next_top,
-					(right as u32).saturating_sub(next_left),
-					(bottom as u32).saturating_sub(next_top),
-				)
-			},
-			FrozenSelectionCorner::TopRight => {
-				let next_right =
-					(right + delta_x).clamp(left.saturating_add(1), monitor_width) as u32;
-				let next_top = (top + delta_y).clamp(0, bottom.saturating_sub(1)) as u32;
-
-				RectPoints::new(
-					left as u32,
-					next_top,
-					next_right.saturating_sub(left as u32),
-					(bottom as u32).saturating_sub(next_top),
-				)
-			},
-			FrozenSelectionCorner::BottomLeft => {
-				let next_left = (left + delta_x).clamp(0, right.saturating_sub(1)) as u32;
-				let next_bottom =
-					(bottom + delta_y).clamp(top.saturating_add(1), monitor_height) as u32;
-
-				RectPoints::new(
-					next_left,
-					top as u32,
-					(right as u32).saturating_sub(next_left),
-					next_bottom.saturating_sub(top as u32),
-				)
-			},
-			FrozenSelectionCorner::BottomRight => {
-				let next_right =
-					(right + delta_x).clamp(left.saturating_add(1), monitor_width) as u32;
-				let next_bottom =
-					(bottom + delta_y).clamp(top.saturating_add(1), monitor_height) as u32;
-
-				RectPoints::new(
-					left as u32,
-					top as u32,
-					next_right.saturating_sub(left as u32),
-					next_bottom.saturating_sub(top as u32),
-				)
-			},
-		}
-	}
-
-	fn apply_frozen_capture_rect_update(
-		&mut self,
-		monitor: MonitorRect,
-		next_rect: RectPoints,
-	) -> bool {
-		if self.state.frozen_capture_rect == Some(next_rect) {
-			return false;
-		}
-
-		self.state.frozen_capture_rect = Some(next_rect);
-
-		let toolbar_default_pos =
-			self.frozen_toolbar_default_position_for_capture_rect(monitor, next_rect);
-		let toolbar_pos = match (
-			self.toolbar_state.floating_position,
-			self.toolbar_state.default_slot_position,
-		) {
-			(Some(floating_pos), Some(default_pos))
-				if !frozen_toolbar_matches_default_slot(floating_pos, default_pos) =>
-			{
-				floating_pos
-			},
-			_ => toolbar_default_pos,
-		};
-
-		self.toolbar_state.default_slot_position = Some(toolbar_default_pos);
-		self.toolbar_state.floating_position = Some(toolbar_pos);
-
-		let should_trace_frozen_selection_drag_timing =
-			self.should_trace_frozen_selection_drag_timing();
-		let toolbar_position_elapsed: Option<Duration> =
-			if should_trace_frozen_selection_drag_timing {
-				let toolbar_position_started_at = Instant::now();
-				let _ = self.update_toolbar_outer_position(monitor, toolbar_pos);
-
-				Some(toolbar_position_started_at.elapsed())
-			} else {
-				let _ = self.update_toolbar_outer_position(monitor, toolbar_pos);
-
-				None
-			};
-		let redraw_request_elapsed: Option<Duration> = if should_trace_frozen_selection_drag_timing
-		{
-			let redraw_request_started_at = Instant::now();
-
-			self.request_redraw_for_monitor(monitor);
-			self.request_redraw_toolbar_window();
-
-			if self.scroll_capture.active {
-				self.request_redraw_scroll_preview_window();
-			}
-
-			Some(redraw_request_started_at.elapsed())
-		} else {
-			self.request_redraw_for_monitor(monitor);
-			self.request_redraw_toolbar_window();
-
-			if self.scroll_capture.active {
-				self.request_redraw_scroll_preview_window();
-			}
-
-			None
-		};
-
-		if should_trace_frozen_selection_drag_timing {
-			tracing::trace!(
-				op = "overlay.frozen_selection_drag.rect_update",
-				monitor_id = monitor.id,
-				rect_x = next_rect.x,
-				rect_y = next_rect.y,
-				rect_width = next_rect.width,
-				rect_height = next_rect.height,
-				toolbar_position_us =
-					toolbar_position_elapsed.map_or(0, |elapsed| elapsed.as_micros()),
-				redraw_request_us = redraw_request_elapsed.map_or(0, |elapsed| elapsed.as_micros()),
-				scroll_capture_active = self.scroll_capture.active,
-				toolbar_outer_pos = ?self.toolbar_outer_pos,
-				toolbar_floating_position = ?self.toolbar_state.floating_position,
-				"Applied frozen selection rect update."
-			);
-		}
-
-		true
-	}
-
-	fn auto_center_frozen_capture_rect(&mut self) -> bool {
-		let Some((monitor, capture_rect)) = self.frozen_capture_rect_drag_target() else {
-			return false;
-		};
-		let Some(capture_image) = self.cropped_frozen_capture_image() else {
-			return false;
-		};
-		let Some(content_bounds) = Self::detect_auto_center_content_bounds(&capture_image) else {
-			return false;
-		};
-		let delta_x_points = Self::auto_center_shift_points(
-			content_bounds.x,
-			content_bounds.width,
-			capture_image.width(),
-			capture_rect.width,
-		);
-		let delta_y_points = Self::auto_center_shift_points(
-			content_bounds.y,
-			content_bounds.height,
-			capture_image.height(),
-			capture_rect.height,
-		);
-		let next_rect = Self::clamp_frozen_capture_rect_to_monitor(
-			monitor,
-			capture_rect.width,
-			capture_rect.height,
-			i64::from(capture_rect.x) + delta_x_points,
-			i64::from(capture_rect.y) + delta_y_points,
-		);
-
-		self.apply_frozen_capture_rect_update(monitor, next_rect)
-	}
-
-	fn auto_center_shift_points(
-		content_origin_px: u32,
-		content_size_px: u32,
-		crop_size_px: u32,
-		capture_size_points: u32,
-	) -> i64 {
-		if crop_size_px == 0 || capture_size_points == 0 {
-			return 0;
-		}
-
-		let content_center_px = content_origin_px as f32 + (content_size_px as f32 * 0.5);
-		let crop_center_px = crop_size_px as f32 * 0.5;
-		let delta_px = content_center_px - crop_center_px;
-
-		((delta_px * capture_size_points as f32) / crop_size_px as f32).round() as i64
-	}
-
-	fn detect_auto_center_content_bounds(image: &RgbaImage) -> Option<RectPoints> {
-		let width = image.width();
-		let height = image.height();
-
-		if width < 2 || height < 2 {
-			return None;
-		}
-
-		let edge_strip = Self::auto_center_edge_strip_extent(width.min(height));
-		let top_mean = Self::region_rgb_mean(image, 0, width, 0, edge_strip)?;
-		let bottom_mean =
-			Self::region_rgb_mean(image, 0, width, height.saturating_sub(edge_strip), height)?;
-		let left_mean = Self::region_rgb_mean(image, 0, edge_strip, 0, height)?;
-		let right_mean =
-			Self::region_rgb_mean(image, width.saturating_sub(edge_strip), width, 0, height)?;
-		let threshold = {
-			let edge_noise = [
-				Self::region_rgb_mean_distance(image, 0, width, 0, edge_strip, top_mean),
-				Self::region_rgb_mean_distance(
-					image,
-					0,
-					width,
-					height.saturating_sub(edge_strip),
-					height,
-					bottom_mean,
-				),
-				Self::region_rgb_mean_distance(image, 0, edge_strip, 0, height, left_mean),
-				Self::region_rgb_mean_distance(
-					image,
-					width.saturating_sub(edge_strip),
-					width,
-					0,
-					height,
-					right_mean,
-				),
-			]
-			.into_iter()
-			.fold(0.0, f32::max);
-
-			(edge_noise * 3.0).round().clamp(24.0, 96.0) as u32
-		};
-		let min_salient_per_row = (width / 64).max(1) as usize;
-		let min_salient_per_column = (height / 64).max(1) as usize;
-		let mut row_counts = vec![0_usize; height as usize];
-		let mut column_counts = vec![0_usize; width as usize];
-
-		for (x, y, pixel) in image.enumerate_pixels() {
-			let salient_distance = [
-				Self::rgb_distance_to_mean(pixel, top_mean),
-				Self::rgb_distance_to_mean(pixel, bottom_mean),
-				Self::rgb_distance_to_mean(pixel, left_mean),
-				Self::rgb_distance_to_mean(pixel, right_mean),
-			]
-			.into_iter()
-			.min()
-			.unwrap_or(0);
-
-			if salient_distance < threshold {
-				continue;
-			}
-
-			row_counts[y as usize] += 1;
-			column_counts[x as usize] += 1;
-		}
-
-		let top = row_counts.iter().position(|count| *count >= min_salient_per_row)?;
-		let bottom = row_counts.iter().rposition(|count| *count >= min_salient_per_row)?;
-		let left = column_counts.iter().position(|count| *count >= min_salient_per_column)?;
-		let right = column_counts.iter().rposition(|count| *count >= min_salient_per_column)?;
-
-		if left > right || top > bottom {
-			return None;
-		}
-
-		let bounds = RectPoints::new(
-			left as u32,
-			top as u32,
-			(right - left + 1) as u32,
-			(bottom - top + 1) as u32,
-		);
-		let fills_crop_width = bounds.width.saturating_mul(100) >= width.saturating_mul(92);
-		let fills_crop_height = bounds.height.saturating_mul(100) >= height.saturating_mul(92);
-
-		if fills_crop_width && fills_crop_height {
-			return None;
-		}
-
-		Some(bounds)
-	}
-
-	fn auto_center_edge_strip_extent(length: u32) -> u32 {
-		((length as f32) * 0.08).round().clamp(1.0, 24.0) as u32
-	}
-
-	fn region_rgb_mean(image: &RgbaImage, x0: u32, x1: u32, y0: u32, y1: u32) -> Option<[f32; 3]> {
-		if x0 >= x1 || y0 >= y1 {
-			return None;
-		}
-
-		let mut r_total = 0_u64;
-		let mut g_total = 0_u64;
-		let mut b_total = 0_u64;
-		let mut sample_count = 0_u64;
-
-		for y in y0..y1 {
-			for x in x0..x1 {
-				let pixel = image.get_pixel(x, y);
-
-				r_total += u64::from(pixel[0]);
-				g_total += u64::from(pixel[1]);
-				b_total += u64::from(pixel[2]);
-				sample_count += 1;
-			}
-		}
-
-		if sample_count == 0 {
-			return None;
-		}
-
-		Some([
-			r_total as f32 / sample_count as f32,
-			g_total as f32 / sample_count as f32,
-			b_total as f32 / sample_count as f32,
-		])
-	}
-
-	fn region_rgb_mean_distance(
-		image: &RgbaImage,
-		x0: u32,
-		x1: u32,
-		y0: u32,
-		y1: u32,
-		mean: [f32; 3],
-	) -> f32 {
-		if x0 >= x1 || y0 >= y1 {
-			return 0.0;
-		}
-
-		let mut total_distance = 0_u64;
-		let mut sample_count = 0_u64;
-
-		for y in y0..y1 {
-			for x in x0..x1 {
-				total_distance +=
-					u64::from(Self::rgb_distance_to_mean(image.get_pixel(x, y), mean));
-				sample_count += 1;
-			}
-		}
-
-		if sample_count == 0 { 0.0 } else { total_distance as f32 / sample_count as f32 }
-	}
-
-	fn rgb_distance_to_mean(pixel: &image::Rgba<u8>, mean: [f32; 3]) -> u32 {
-		(pixel[0] as f32 - mean[0]).abs().round() as u32
-			+ (pixel[1] as f32 - mean[1]).abs().round() as u32
-			+ (pixel[2] as f32 - mean[2]).abs().round() as u32
 	}
 
 	fn cropped_frozen_capture_image(&self) -> Option<RgbaImage> {
